examples/llamaindex-python/requirements.txt
llama-index-core>=0.11.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
